resources: split requests resource and config
